Move types back to internal.
