+++ pkg/river/internal/rivertags/rivertags_test.go
@@ -4,7 +4,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/grafana/agent/pkg/river/rivertags"
+	"github.com/grafana/agent/pkg/river/internal/rivertags"
 
 	"github.com/stretchr/testify/require"
 )
