+++ pkg/river/internal/value/decode_benchmarks_test.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/grafana/agent/pkg/river/value"
+	"github.com/grafana/agent/pkg/river/internal/value"
 )
 
 func BenchmarkObjectDecode(b *testing.B) {
