+++ pkg/river/internal/value/decode_test.go
@@ -7,7 +7,7 @@ import (
 	"time"
 	"unsafe"
 
-	"github.com/grafana/agent/pkg/river/value"
+	"github.com/grafana/agent/pkg/river/internal/value"
 	"github.com/stretchr/testify/require"
 )
