+++ pkg/river/internal/value/number_value.go
