+++ pkg/river/internal/value/tag_cache.go
@@ -3,7 +3,7 @@ package value
 import (
 	"reflect"
 
-	"github.com/grafana/agent/pkg/river/rivertags"
+	"github.com/grafana/agent/pkg/river/internal/rivertags"
 )
 
 // tagsCache caches the river tags for a struct type. This is never cleared,
