+++ pkg/river/internal/value/value_object_test.go
@@ -3,7 +3,7 @@ package value_test
 import (
 	"testing"
 
-	"github.com/grafana/agent/pkg/river/value"
+	"github.com/grafana/agent/pkg/river/internal/value"
 	"github.com/stretchr/testify/require"
 )
