+++ pkg/river/token/builder/builder.go
@@ -9,7 +9,7 @@ import (
 	"reflect"
 	"strings"
 
-	"github.com/grafana/agent/pkg/river/rivertags"
+	"github.com/grafana/agent/pkg/river/internal/rivertags"
 
 	"github.com/grafana/agent/pkg/river/token"
 )
